add support for native_library_merge_map
Summary:
The application of the merge map is really straightforward, the main
complexity here is in implementing the handling of producing the link args for
the merged libs.

There's a lot of complexity in inconsistencies in buck1 around how this merging
is implemented and how static libs are handled, how linker flags and exported
linker flags are handled. I've tried to produce a more internally consistent
approach to this that still is close enough to buck1 that we don't need to
make significant changes.

One particular difference that we see when enabling this is that there is less
implicit exclusions from the merge process. This means that we need more explicit
exclusions.

Reviewed By: mzlee

Differential Revision: D49141125

fbshipit-source-id: 42ba75804ac9256a630d4b472c9b22cdbe36b776