Correctly read `cxx_toolchain_override.shared_library_interface_mode`
Summary:
**Problem**

The `shlib_interfaces` field of `LinkerInfo` was being set to a `string` for the Apple Xcode `cxx_toolchain()`. The code in question:

https://www.internalfb.com/code/fbsource/[bdd359ac67d21a514551fbb44ec43baa25259da4]/fbcode/buck2/prelude/cxx/user/cxx_toolchain_override.bzl?lines=94

The type of `ctx.attrs.shared_library_interface_mode` is a `[string, None]` while `base_linker_info.shlib_interfaces` is of type `ShlibInterfacesMode`.

Consequently, when creating the Xcode toolchain, we ended up passing a string:

https://www.internalfb.com/code/fbsource/[bdd359ac67d21a514551fbb44ec43baa25259da4]/xplat/buck2/platform/apple/create_apple_toolchain.bzl?lines=140

which means the type of `LinkInfo.shlib_interfaces` becomes a `string` rather than `ShlibInterfacesMode`. This meant that the check whether to enable shlib interfaces **always passed** because it was comparing the string `disabled` to the enum value `ShlibInterfacesMode("disabled")`

https://www.internalfb.com/code/fbsource/[bdd359ac67d21a514551fbb44ec43baa25259da4]/fbcode/buck2/prelude/cxx/cxx_library_utility.bzl?lines=152

This meant that we couldn't disable shlib interfaces for Apple which is broken atm.

## Typing

D49536198 adds the requisite typing that would have caught this. It's a separate diff as it detects another bug in the fbcode toolchains.

Reviewed By: krallin

Differential Revision: D49536195

fbshipit-source-id: 7e288374b3844bb2b628c351c5e94164807cb802