add transitive dependencies for C# and .net dll references
Summary:
Adds transitive dependencies to the `csharp_library` and `prebuilt_dotnet_library` rules to ensure the dependencies of downstream targets are correctly listed as arguments to `csc.exe`.

This was tested with a new test case where `A` depends on `B` depends on `C` (so no direct depenedncy from `A` to `C`), and where `A` calls a method exported from `C`. Without tsets this fails, but with the new tset implementation the test passes.

Reviewed By: akrieger

Differential Revision: D49613504

fbshipit-source-id: 21e396ab8e470edb01b4086679dbb9ffb558b6f3