apple_bundle() builds all genrule() deps
Summary:
**Context**

In Buck1, it's common to have `genrule()` deps on `apple_bundle()` which get run as part of a build. Buck2 breaks the behavior and any such `genrule()` deps will not be run.

**Changes**

Adds three ways to control the behavior of building genrule deps, in order of priority:

- `build_genrule_deps` field on targets
- `apple.build_genrule_deps` global buckconfig
- `config//features/apple/constraints:build_genrule_deps_enabled` constraint

If any of these is turned on, `genrule()` deps of `apple_bundle()` targets become hidden inputs to the bundling command, thus will be built as part of bundling.

Reviewed By: blackm00n

Differential Revision: D49879350

fbshipit-source-id: 084f07482190d6773a0bdbadf31e685825237615