Mark all single split-debug builds as having external debug info
Summary:
ThinLTO + split-dwarf produces the majority of debug info in a dwo dir,
but since we disable thinlto for some compiled objects, those CUs are
compiled directly into object code with embedded dwo info instead of
emitting dwo debug info at link time.

Ideally, we should track whether a CU is compiled to bitcode or not
AND whether it will participate in thinLTO, but currently we only have
information on the linker. This diff unblocks lto builds in the mean time.

Differential Revision: D50138571

fbshipit-source-id: 6f74f5cb9e8ec8c855e2a6fd215ec5848421d7a4