Revert D49974422: Multisect successfully blamed "D49974422: [buck2][python] use anon targets for compilation" for test or build failures
Summary:
This diff is reverting D49974422
Depends on D50401645
D49974422: [buck2][python] use anon targets for compilation by BrandonTheBuilder has been identified to be causing the following test or build failures:

Tests affected:
- [registry/builder/fbpkg/tests:test-caf-builder-e2e - test_e2e (registry.builder.fbpkg.tests.test_caf_builder_e2e.CafBuilderE2ETestCase)](https://www.internalfb.com/intern/test/281475088848239/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3330598
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: lmvasquezg

Differential Revision: D50401655

fbshipit-source-id: 10aa079de61d1738057b233a073479857b2a8dc0