Support executable_name parameter on cxx_binary
Summary:
WTTS.

This is slightly convoluted because python rules can call into `cxx_executable` and also have
an `executable_name` parameter, so reading off `ctx.attrs` directly would change the output
that the python rules depend on and breaks things. Instead, we sideload the `executable_name`
into CxxRuleConstructorParams which is actually only instantiated in cxx rules, so that when
we read it back in `_link_into_executable` the value is only set in direct calls to `cxx_binary`.

Not changing the public `get_cxx_executable_product_name` function mildly concerns me but if
tests are green I won't worry about it.

Reviewed By: milend

Differential Revision: D51121760

fbshipit-source-id: 2e5c8cf92bd6c220e84f02aeba0c1ea5b2a1366f