toolchain: Set up no-sysroot deps toolchain
Summary:
This diff lays out how I want to avoid the cyclic dependency issue with explicit sysroot deps compiled from source.

Doing this by defining a second toolchain is generally appropriate I think. Asking crates that expect to be compiled as members of the sysroot to declare that by specifying a non-default toolchain seems correct. Also, a constraint would be inappropriate here because we do not generally want to propagate this to all target dependencies, and because it runs into risks of accidentally ending up with the same dep in your build twice, under two different configurations.

I didn't bother trying to add support to `override_rust_toolchain` for this because we'd need something like `attrs.struct()` to make that not super ugly. Plus, rules are good and people should write more of them.

The one potentially controversial decision in this diff is the choice to put the override above the massive toolchain select tree, instead of below. On balance, I'd guess that this is the better thing to do. This override is only responsible for inserting the sysroot targets; I expect that it'll only ever grow two cases, one for fbcode sysroot targets and one for fbsource ones. That means we'll never need all the complicated machinery that's in the big select tree. Plus, selectified toolchains are good anyway.

Reviewed By: capickett

Differential Revision: D51014702

fbshipit-source-id: 970de5291aa725e1e5e1e22fdde53603e1e4d412