fix resource brokers duplicates
Summary: We don't want `apple_simulators` target to be configured differently and handled as a different resource broker by buck2 core. Since `configured_alias` target is not a single node in a graph (but forwards to a single node when being configured with different platforms) change the approach in how we setup the simulator targets and use a transition + transition_dep to achieve the desired outcome.

Reviewed By: milend

Differential Revision: D51303334

fbshipit-source-id: 424fe94f8d180fe52301e49e2585a0c64d56f08c