Support group entries in native_library_merge_sequence
Summary:
## This stack
We're adding support for "group entries" in `native_library_merge_sequence`. These will contain *multiple* definitions for libraries and treat them all as having the same priority.

Previously, merged libraries with interchangeable priority had to be ordered arbitrarily, which could force large libraries to depend on one another, adding significant unnecessary overhead to the loading latency of the dependent library. If such libraries are part of the same group entry, their shared target dependencies will instead be moved to a separate merged library, resulting in much stabler latency outcomes. (Note that this behavior is *not* optimal if libraries do have identifiably distinct priorities. It should be used only when finely-tuned configuration is not practical.)

## This diff

And here is Buck 2. [I expected that I would not be able to handle this](https://fb.workplace.com/groups/691614554220496/posts/25433389379616337/?comment_id=25433412022947406&reply_comment_id=25434023196219622), but the structure of the Python script ended up working quite well for group entries; it was really only the details of the logic that needed to be changed significantly.

NOTE: **This is not integrated into Bazel code** - it's missing the equivalent of D47902651. D51567455 is a severe hack that makes the new argument syntax testable but totally type-unsafe.

Reviewed By: cjhopman

Differential Revision: D51567454

fbshipit-source-id: 78c67e591a1739f6232cd3533d8ef2a54f8cd127