Unpack ExecutableSharedLibArguments one function later
Summary:
Later in this stack, D51222268 is going to add another field `external_debug_info` to the `ExecutableSharedLibArguments` record, which needs to get propagated through this code.

Instead of needing to grow a 2-tuple into a 3-tuple in that diff, I am replacing the return type here with a record.

Reviewed By: podtserkovskiy

Differential Revision: D51742572

fbshipit-source-id: 289676f127a9b7330eaea564a9f5aae120b3ba45