Separately track Go runtime files and other outputs
Summary:
The `go_binary` rule builds `DefaultInfo` and `DistInfo` by unioning together the `other_outputs` and `runtime_files` from a collection of resources that it depends on.

Previously, it built just one list that was reused across both `DefaultInfo.other_outputs` and `DistInfo.nondebug_runtime_files`. This diff changes that to separate lists. The effect is that if a Go binary depends on a resource (such as a C++ executable) that has external debuginfo, the external debuginfo is *not* treated as a mandatory runtime dependency of the Go binary. It is only materialized if the Go binary is the output of a build, not when the Go binary is depended on by another rule.

The summary of {D51222272} describes the distinction in more detail.

Reviewed By: podtserkovskiy

Differential Revision: D51222271

fbshipit-source-id: 81ed809e9f1f83381ae7f7ccfc461ee22ab477ef