Force declared outputs of files to live in different directories
Summary:
Occasionally, a js_library action running on remote execution has some trouble writing the final output. The action gets a permission denied error, because the buck-out directory it writes to is owned by root or a system user, with chmod 755 permissions, meaning the non-root user running the action is not allowed to write there.

My current working theory is that the issue lies in the reuse of the same buck-out directories across multiple rules. For example. library_dependencies folders take in library_files action outputs, but they're using the same base directories. It could be the case that there's some sort of race copying the input files in and creating the declared output dir, where if the copying

To try and get around this issue, I'm modifying the paths of the declared outputs to live in their own directories.

Reviewed By: motiz88

Differential Revision: D51982209

fbshipit-source-id: 3692c27f1be877ddc111bdd0f5c59c02b5a6ea6f