Add toolchain_alias as a rule
Summary: If you want to do anything cross compilation or anything, you pretty much have to introduce a toolchain alias. If Buck1 had toolchains, this would definitely be part of the prelude, and available by default. I appreciate we don't want to add new things to the prelude, but given how fundamental this is, I think we should do it.

Reviewed By: JakobDegen

Differential Revision: D51535366

fbshipit-source-id: 0094c0400a910abca2841529d661bf17c3d18a4a