Default produce_interface_from_stub_shared_library to False
Summary:
This mode doesn't work correctly right now because it doesn't account
for static libraries that are included in the shared library interface.
https://fb.workplace.com/groups/buck2users/permalink/3551139855142370/
has more details.

From what I understand, the purpose of this mode is to increase link
parallelism, such that you can link a target without waiting for all its
shared library dependencies to be linked first. I don't think it's
possible to accomplish that in a sound manner though (sound as in the
interface libraries reflect the actual libraries exactly). Consider
something like the following:

```
lib1.a defines the function foo
lib2.so links against lib1.a and so exports foo
lib3.so links against both lib2.so and lib1.a in that order
```

In the case where lib3.so only references foo and no other symbols from
lib1.a (technically whichever object file inside lib1.a defines foo),
and lib1.a is not `link_whole = True`, lib3.so will end up referencing
foo from lib2.so instead of exporting its own copy. For producing the
stub shared library for lib3.so, we have two options:
- We could link against only static library dependencies, which would
  fix the bug in my post, but then the stub library would incorrectly
  export foo.
- We could link against static library dependencies as well as the stub
  libraries corresponding to the shared library dependencies, in the
  exact same order as the actual link, but then we've lost the link
  parallelism so there's no advantage to this mode.

If my reasoning is correct, we should just get rid of this mode entirely
IMO, because I don't think the increased link parallelism is worth the
potential correctness footguns.

Reviewed By: andrewjcg

Differential Revision: D51926025

fbshipit-source-id: 205652efc97b147ba577e27977da1bfcf43c7dae